+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### IGCIT Helper v1.2.0
+* Add device name and manufacturer
+* Add memory information
+* Add copy button to all values
+* Add menu option to take screenshot of Helper window
+
 #### IGCIT Driver Switch v1.0.0
 * Initial release